+++ tests/pipelines/misc/test_dates.py
@@ -6,7 +6,7 @@ import spacy
 from pytest import fixture
 from spacy.language import Language
 
-from edsnlp.pipelines.misc.dates.models import AbsoluteDate
+from edsnlp.pipelines.misc.dates.models import AbsoluteDate, Relative
 from edsnlp.utils.examples import parse_example
 
 TZ = pytz.timezone("Europe/Paris")
@@ -88,7 +88,7 @@ def test_dates_component(blank_nlp: Language):
         assert len(spans) == len(entities)
         assert len(doc.ents) == len(entities)
 
-        for span, entity in zip(doc.spans["dates"], entities):
+        for span, entity in zip(spans, entities):
             assert span.text == text[entity.start_char : entity.end_char]
 
             date = span._.date if span.label_ == "date" else span._.duration
@@ -145,8 +145,11 @@ def test_dates_component(blank_nlp: Language):
                         note_datetime=note_datetime, infer_from_context=True
                     ) == TZ.localize(datetime(**d))
 
+            elif isinstance(date, Relative):
+                assert date.to_datetime() is None
             else:
-                assert date.to_datetime() is not None
+                assert date.to_duration()
+                assert date.to_datetime(note_datetime=note_datetime)
 
 
 def test_periods(blank_nlp: Language):
@@ -186,9 +189,11 @@ def test_time(with_time: bool):
 
         doc = nlp(text)
 
-        assert len(doc.spans["dates"]) == len(entities)
+        spans = sorted(doc.spans["dates"] + doc.spans["durations"])
 
-        for span, entity in zip(doc.spans["dates"], entities):
+        assert len(spans) == len(entities)
+
+        for span, entity in zip(spans, entities):
             assert span.text == text[entity.start_char : entity.end_char]
             norm = next(m.value for m in entity.modifiers if m.key == "norm")
             assert span._.date.norm() == norm
@@ -216,7 +221,7 @@ def test_false_positives(blank_nlp: Language):
     for example in counter_examples:
         doc = blank_nlp(example)
 
-        assert len(doc.spans["dates"]) == 0
+        assert len((*doc.spans["dates"], *doc.spans["durations"])) == 0
 
 
 def test_dates_on_ents_only():
@@ -239,9 +244,11 @@ def test_dates_on_ents_only():
 
     assert len(doc.ents) == 1
 
-    assert len(doc.spans["dates"]) == len(entities)
+    spans = sorted(doc.spans["dates"] + doc.spans["durations"])
 
-    for span, entity in zip(doc.spans["dates"], entities):
+    assert len(spans) == len(entities)
+
+    for span, entity in zip(spans, entities):
         assert span.text == text[entity.start_char : entity.end_char]
 
 
@@ -252,5 +259,5 @@ def test_illegal_dates(blank_nlp):
     )
     for text in texts:
         doc = blank_nlp(text)
-        ent = doc.spans["dates"][0]
+        ent = sorted((*doc.spans["dates"], *doc.spans["durations"]))[0]
         assert ent._.date.to_datetime() is None